cleanup: add missing rss button & add margin only to the first child
Till now, all buttons inside the card header were getting a margin of 30px, but that adds a lot of space not nessesary for other than the first button. So I've made the 30px marging only be added to the first child button of the card's header.

Sorry for the wall of text, but I wanted to explain exactly why I did this change.